Adding memcached support to php7.
Using libmemcached version 1.0.18, as linked from the PHP extension
page.  There's a bit of cheating here, to build the lib without the
clients and the (broken due to lack of `alloca.h`) tests.  It saves a
lot of time.

Due to libmemcached being written in C++, PHP's link step needs to use
`$(CXX)` and we have to define that for `configure`, to solve the lack
of exception handling symbols (contrary to reports on the internet, this
was _not_ fixed by including `-lstdc++` after `-lmemcached` options.)

This could be the wrong way to do things, but it's _a_ way.