Patch and enable opcache.
As usual, this **compiles,** but I have yet to run and test it.

The distributed `config.m4` does not work in cross-compile mode.

For the SHM tests, instead of running the programs on the build machine,
we just link them.  The actual code using the results tries to use every
one at runtime that was defined as working at configure time, and uses
the first one that seems to work.

Compile time flock struct test based on a program from Antti Kantee.

The header file that _uses_ the result of that check only takes the
results as a fallback, in case it doesn't recognize the specific
BSD/Linux/etc. platform defines, but both that file and the toolchain
support `__NetBSD__`.  But it's a cool hack, so I left it in.